reorg silentsyntax, comments, more annotations, revert accidents
