+++ projects/batfish/src/test/resources/org/batfish/main/testconfigs/annotate-hierarchical-before
@@ -23,3 +23,4 @@ protocols {
     }
   }
 }
+set routing-options resolution rib inet.4 import policy1
